described mod spend_wallet and bumped it to a new file
